+++ src/services/test/socket.service.spec.ts
@@ -5,12 +5,12 @@ import * as sinon from 'sinon';
 import * as sinonChai from 'sinon-chai';
 use(sinonChai);
 
-import { MiterSocketService } from '../socket.service';
+import { SocketService } from '../socket.service';
 
-describe('MiterSocketService', () => {
-    it('should have a boilerplate test', () => {
+describe('SocketService', () => {
+    it('should have a placeholder test', () => {
         expect(true).to.be.true;
     });
     
-    //TODO: test MiterSocketService
+    //TODO: test SocketService
 });